opentv: fix long event duration duration limit (65535) is not large enough, 20 hour events (72000 seconds) can be seen.

+++ lib/dvb/opentv.cpp
@@ -224,7 +224,7 @@ uint16_t OpenTvTitle::getEventId(void) const
 	return eventId;
 }
 
-uint16_t OpenTvTitle::getDuration(void) const
+uint32_t OpenTvTitle::getDuration(void) const
 {
 	return duration;
 }
